Add initial version
This API is written in Python and uses CherryPy as base framework. The
decision to use such a low level framework was to get control over raw
request (e.g. file transfer).

Low level interventions are done via different CherryPy tools. They
begin with 'cherry_*'.

The documentation is in written in asciidoc format and can be exported
to several destination formats.

To create a clean build start the following command:

  ./waf configure build install

This will install all required files into the 'out' folder in the root
of this repository.